ref: move interceptors logic to core file, refactor spinner component
