+++ src/core/LoaderSpinner.vue
@@ -5,62 +5,22 @@
 </template>
 
 <script lang="ts">
-import { defineComponent, ref } from 'vue';
+import { defineComponent } from 'vue';
 import ProgressSpinner from 'primevue/progressspinner';
-import { HTTP, HTTP_AUTH } from '@/core/api/http-common';
-import { AxiosError, AxiosRequestConfig, AxiosResponse } from 'axios';
 
 export default defineComponent({
   name: 'LoaderSpinner',
   components: {
     ProgressSpinner,
   },
-
+  props: {
+    isLoading: {
+      type: Boolean,
+      required: true,
+    },
+  },
   setup() {
-    const isLoading = ref(false);
-
-    HTTP.interceptors.request.use(
-      (req: AxiosRequestConfig) => {
-        isLoading.value = true;
-        return req;
-      },
-      (error: AxiosError) => {
-        isLoading.value = false;
-        return error;
-      }
-    );
-    HTTP.interceptors.response.use(
-      (response: AxiosResponse) => {
-        isLoading.value = false;
-        return response;
-      },
-      (error: AxiosError) => {
-        isLoading.value = false;
-        return error;
-      }
-    );
-    HTTP_AUTH.interceptors.request.use(
-      (req: AxiosRequestConfig) => {
-        isLoading.value = true;
-        return req;
-      },
-      (error: AxiosError) => {
-        isLoading.value = false;
-        return error;
-      }
-    );
-    HTTP_AUTH.interceptors.response.use(
-      (response: AxiosResponse) => {
-        isLoading.value = false;
-        return response;
-      },
-      (error: AxiosError) => {
-        isLoading.value = false;
-        return error;
-      }
-    );
-
-    return { isLoading };
+    return {};
   },
 });
 </script>
